requirements.txt
Django~=4.0.6
paho_mqtt~=1.6.1
pymongo~=4.2.0
django-cors-headers~=3.13.0